+++ lib/trailblazer/developer.rb
@@ -9,6 +9,7 @@ end
 require "trailblazer/activity"
 require "trailblazer/developer/wtf"
 require "trailblazer/developer/wtf/renderer"
+require "trailblazer/developer/trace/snapshot"
 require "trailblazer/developer/trace"
 require "trailblazer/developer/trace/stack"
 require "trailblazer/developer/trace/tree"