+++ test/trace_test.rb
@@ -103,9 +103,6 @@ class TraceTest < Minitest::Spec
 `-- End.success}
   end
 
-  require "trailblazer/developer/trace/snapshot"
-
-
   it "nested tracing with better-snapshot" do
     # Test custom classes without explicit {#hash} implementation.
     class User
@@ -154,14 +151,18 @@ class TraceTest < Minitest::Spec
       end
     end
 
-    inspect_only_flow_options = {}
-
     Snapshot = Trailblazer::Developer::Trace::Snapshot
 
+    inspect_only_flow_options = {
+      input_data_collector:   Snapshot::Deprecated.method(:default_input_data_collector),
+      output_data_collector:  Snapshot::Deprecated.method(:default_output_data_collector),
+    }
+
+
     snapshot_flow_options = {
-      input_data_collector:   Snapshot.method(:input_data_collector),
-      output_data_collector:  Snapshot.method(:output_data_collector),
-      variable_versions:      Snapshot::Versions.new
+      input_data_collector:   Snapshot.method(:before_snapshooter),
+      output_data_collector:  Snapshot.method(:after_snapshooter),
+      variable_versions:      Snapshot::Ctx::Versions.new
     }
 
     activity = namespace::Endpoint
@@ -196,9 +197,9 @@ class TraceTest < Minitest::Spec
 
 
     snapshot_flow_options = {
-      input_data_collector:   Snapshot.method(:input_data_collector),
-      output_data_collector:  Snapshot.method(:output_data_collector),
-      variable_versions:      Snapshot::Versions.new
+      input_data_collector:   Snapshot.method(:before_snapshooter),
+      output_data_collector:  Snapshot.method(:after_snapshooter),
+      variable_versions:      Snapshot::Ctx::Versions.new
     }
 
     stack, signal, (ctx, flow_options) = Dev::Trace.invoke(